Added logic to mute all sounds
